+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## [Unreleased]
 
 ### Added
--  Apache Flink 1.16
+
+-   Apache Flink 1.16
 
 ### Fixed
--  fixed dependencies version in requirements.txt
+
+-   Jar paths in `pipeline.jars` property are parsed and passed when creating remote execution environment.
+-   fixed dependencies version in requirements.txt
 
 ## [0.12.1] - 2023-01-05